+++ FS_Server/src/main/java/com/bonree/brfs/duplication/FidBuilder.java
@@ -4,7 +4,7 @@ import com.bonree.brfs.duplication.coordinator.FileNode;
 
 public class FidBuilder {
 	
-	public static String getFid(FileNode node, int offset, int size) {
+	public static String getFid(FileNode node, long offset, int size) {
 		StringBuilder builder = new StringBuilder();
 		builder.append(node.getStorageName()).append("-").append(node.getName()).append("-")
 		       .append(offset).append("-").append(size);
